create servlet for comments section
